Modify API for Cortex
